requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.2.0=eigen
absl-py=1.4.0=pyhd8ed1ab_0
aiohttp=3.8.1=py37hcc03f2d_1
aiosignal=1.3.1=pyhd8ed1ab_0
anyio=3.6.2=pyhd8ed1ab_0
astor=0.8.1=pyh9f0ad1d_0
astunparse=1.6.3=pyhd8ed1ab_0
async-timeout=4.0.2=pyhd8ed1ab_0
asynctest=0.13.0=py_0
attrs=23.1.0=pyh71513ae_1
blinker=1.6.2=pyhd8ed1ab_0
brotlipy=0.7.0=py37hcc03f2d_1004
ca-certificates=2023.5.7=h56e8100_0
cachetools=5.3.0=pyhd8ed1ab_0
certifi=2023.5.7=pyhd8ed1ab_0
cffi=1.15.1=py37ha95fbe2_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.3=py37h03978a9_0
colorama=0.4.6=pyhd8ed1ab_0
cryptography=3.4.8=py37h65266a2_1
fastapi=0.95.2=pyhd8ed1ab_0
gunicorn==20.0.4
flatbuffers=23.3.3=h63175ca_0
freetype=2.12.1=h546665d_1
frozenlist=1.3.3=py37h2bbff1b_0
gast=0.4.0=pyh9f0ad1d_0
giflib=5.2.1=h64bf75a_3
google-auth=2.18.1=pyh1a96a4e_0
google-auth-oauthlib=0.4.1=py_1
google-pasta=0.2.0=pyh8c360ce_0
grpcio=1.42.0=py37hc60d5dd_0
h11=0.14.0=pyhd8ed1ab_0
h5py=2.10.0=nompi_py37he280515_106
hdf5=1.10.6=nompi_h5268f04_1114
icu=58.2=ha925a31_3
idna=3.4=pyhd8ed1ab_0
importlib-metadata=4.11.4=py37h03978a9_0
intel-openmp=2023.1.0=h57928b3_46319
jbig=2.1=h8d14728_2003
jpeg=9e=hcfcfb64_3
keras=2.10.0=py37haa95532_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd8ed1ab_0
krb5=1.20.1=h6609f42_0
lcms2=2.12=h83e58a3_0
lerc=3.0=hd77b12b_0
libblas=3.9.0=16_win64_mkl
libcblas=3.9.0=16_win64_mkl
libcurl=7.88.1=h86230a5_0
libdeflate=1.8=h2bbff1b_5
libhwloc=2.9.1=h51c2c0f_0
libiconv=1.17=h8ffe710_0
liblapack=3.9.0=16_win64_mkl
libpng=1.6.39=h19919ed_0
libprotobuf=3.20.3=h12be248_0
libsqlite=3.42.0=hcfcfb64_0
libssh2=1.10.0=h680486a_3
libtiff=4.3.0=hd413186_2
libxml2=2.10.4=hc3477c8_0
libzlib=1.2.13=hcfcfb64_4
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.4.3=pyhd8ed1ab_0
mkl=2022.1.0=h6a75c08_874
msys2-conda-epoch=20160418=1
multidict=6.0.2=py37h2bbff1b_0
numpy=1.21.6=py37h2830a78_0
oauthlib=3.2.2=pyhd8ed1ab_0
olefile=0.46=pyh9f0ad1d_1
openjpeg=2.5.0=hb211442_0
openssl=1.1.1t=hcfcfb64_0
opt_einsum=3.3.0=pyhd8ed1ab_1
packaging=23.1=pyhd8ed1ab_0
pillow=8.4.0=py37hd7d9ad0_0
pip=23.1.2=pyhd8ed1ab_0
protobuf=3.20.3=py37hd77b12b_0
pthreads-win32=2.9.1=hfa6e2cd_3
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd8ed1ab_0
pydantic=1.10.2=py37h2bbff1b_0
pyjwt=2.7.0=pyhd8ed1ab_0
pyopenssl=20.0.1=pyhd8ed1ab_0
pyreadline=2.1=py37h03978a9_1006
pysocks=1.7.1=py37h03978a9_5
python=3.7.12=h7840368_100_cpython
python-flatbuffers=23.1.21=pyhd8ed1ab_0
python-multipart=0.0.6=pyhd8ed1ab_0
python_abi=3.7=3_cp37m
pyu2f=0.1.5=pyhd8ed1ab_0
requests=2.31.0=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
scipy=1.7.3=py37hb6553fb_0
setuptools=67.7.2=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.10=hfb803bf_0
sniffio=1.3.0=pyhd8ed1ab_0
sqlite=3.42.0=hcfcfb64_0
starlette=0.27.0=pyhd8ed1ab_0
tbb=2021.9.0=h91493d7_0
tensorboard=2.10.0=py37haa95532_0
tensorboard-data-server=0.6.1=py37haa95532_0
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
tensorflow=2.10.0=eigen_py37h0b514e4_0
tensorflow-base=2.10.0=eigen_py37he3c91d7_0
tensorflow-estimator=2.10.0=py37haa95532_0
termcolor=2.3.0=pyhd8ed1ab_0
tk=8.6.12=h8ffe710_0
typing-extensions=4.6.1=hd8ed1ab_0
typing_extensions=4.6.1=pyha770c72_0
ucrt=10.0.22621.0=h57928b3_0
urllib3=1.26.15=pyhd8ed1ab_0
uvicorn=0.19.0=py37h03978a9_0
vc=14.3=hb25d44b_16
vc14_runtime=14.34.31931=h5081d32_16
vs2015_runtime=14.34.31931=hed1258a_16
werkzeug=2.1.2=pyhd8ed1ab_1
wheel=0.40.0=pyhd8ed1ab_0
win_inet_pton=1.1.0=pyhd8ed1ab_6
wrapt=1.11.2=py37h4ab8f01_1
xz=5.2.6=h8d14728_0
yarl=1.7.2=py37hcc03f2d_2
zipp=3.15.0=pyhd8ed1ab_0
zlib=1.2.13=hcfcfb64_4
zstd=1.5.2=h12be248_6
